+++ src/commands/models/list.format.test.ts
@@ -1,7 +1,28 @@
 // Model list formatting tests cover fixed-width terminal cell helpers.
 import { describe, expect, it } from "vitest";
 import { visibleWidth } from "../../../packages/terminal-core/src/ansi.js";
-import { pad, truncate } from "./list.format.js";
+import { formatTokenK, pad, truncate } from "./list.format.js";
+
+describe("formatTokenK", () => {
+  it("renders token context windows in decimal K", () => {
+    expect(formatTokenK(200_000)).toBe("200k");
+    expect(formatTokenK(128_000)).toBe("128k");
+    expect(formatTokenK(1_000_000)).toBe("1000k");
+    expect(formatTokenK(195_000)).toBe("195k");
+  });
+
+  it("passes small counts through and switches to K at 1000", () => {
+    expect(formatTokenK(999)).toBe("999");
+    expect(formatTokenK(1_000)).toBe("1k");
+  });
+
+  it("returns a dash for missing or non-finite values", () => {
+    expect(formatTokenK(undefined)).toBe("-");
+    expect(formatTokenK(null)).toBe("-");
+    expect(formatTokenK(0)).toBe("-");
+    expect(formatTokenK(Number.NaN)).toBe("-");
+  });
+});
 
 describe("truncate", () => {
   it("preserves existing ASCII truncation with an ellipsis suffix", () => {
