+++ src/commands/models/list.registry-load.ts
@@ -1,3 +1,4 @@
+import { modelKey } from "../../agents/model-ref-shared.js";
 import { shouldSuppressBuiltInModel } from "../../agents/model-suppression.js";
 /** Registry-loading adapters for model-list row construction. */
 import { loadPreparedAgentModelRegistry as loadAgentModelRegistry } from "../../agents/prepared-model-registry.js";
@@ -6,7 +7,6 @@ import type { ModelRegistry } from "../../llm/model-registry.js";
 import type { Model } from "../../llm/types.js";
 import { loadModelRegistry } from "./list.registry.js";
 import type { ConfiguredEntry } from "./list.types.js";
-import { modelKey } from "./shared.js";
 
 /** Loads the full model registry and tracks discovered provider/model keys. */
 export async function loadListModelRegistry(
